implement iocp and overlapped io model

+++ src/main/java/com/github/jeffreystoke/winsock/io/constant/SocketError.kt
@@ -1,6 +1,6 @@
 package com.github.jeffreystoke.winsock.io.constant
 
-enum class SocketError(val value: Int) {
+enum class SocketError(val value: Long) {
     INVALID_SOCKET(0.inv()),
     Error(-1),
 }
